+++ src/Database/Seeders/IgUserSeeder.php
@@ -2,7 +2,6 @@
 
 namespace InternetGuru\LaravelUser\Database\Seeders;
 
-use InternetGuru\LaravelUser\Enums\Provider;
 use App\Models\User;
 use Illuminate\Database\Seeder;
 
@@ -19,7 +18,6 @@ class IgUserSeeder extends Seeder
             User::providers()::GOOGLE,
             '106988997789517413087',
             'George Pavelka',
-            'pavelka.iix@gmail.com',
         )->create([
             'name' => 'George',
             'email' => 'george@giftcarder.io',
@@ -30,12 +28,10 @@ class IgUserSeeder extends Seeder
             User::providers()::SEZNAM,
             'e273eac080e7aa2ea6158e56620ed3df7ca1c8d1',
             'Pavel Petržela',
-            'poulikpetrzela@seznam.cz',
         )->withSocialite(
             User::providers()::GOOGLE,
             '108829846781865836650',
             'Pavel Petržela',
-            'poulikpetrzela@gmail.com',
         )->create([
             'name' => 'Pavel',
             'email' => 'pavel@giftcarder.io',